fix: reject maxFill above type(uint128).max instead of saturating the budget
The L-2 mitigation saturated the remaining budget at type(uint128).max
before calling ICallbackFeeAdjuster.beforeDispatch(), so the router was
still making a sizing decision on the adjuster's behalf: an arbitrary
adjuster may size type(uint128).max + 1 differently than
type(uint128).max. Per the auditor's follow-up, revert upfront with
MaxFillTooLarge when maxFill (post sentinel resolution) exceeds
type(uint128).max: remaining <= maxFill then always fits uint128, the
adjuster sees the exact budget, and the WAD inversions stay
overflow-free. The bound is not a hard Midnight limit but is sufficient
for all practical purposes; the adapter's uint256.max sentinel resolves
before the check.

+++ test/integration/MixedExecutionDCOFIntegration.t.sol
@@ -211,7 +211,7 @@ contract MixedExecutionDCOFIntegrationTest is MigrationRatifierTestBase {
         return ExecuteParams({
             deadline: 0,
             fillAxis: FillAxis.UNITS,
-            maxFill: type(uint256).max,
+            maxFill: type(uint128).max,
             minFill: 0,
             minPrice: 0,
             maxPrice: type(uint256).max,
